refactor: Rename all HelloWorld components
