+++ README.md
@@ -128,3 +128,8 @@ sam delete --stack-name "sam-app"
 See the [AWS SAM developer guide](https://docs.aws.amazon.com/serverless-application-model/latest/developerguide/what-is-sam.html) for an introduction to SAM specification, the SAM CLI, and serverless application concepts.
 
 Next, you can use AWS Serverless Application Repository to deploy ready to use Apps that go beyond hello world samples and learn how authors developed their applications: [AWS Serverless Application Repository main page](https://aws.amazon.com/serverless/serverlessrepo/)
+
+
+## Project Roadmap
+
+You can find the detailed project plan, milestones, and technical specifications in the [full roadmap document](./docs/roadmap.md).
