+++ template.yaml
@@ -17,10 +17,10 @@ Globals:
    LoggingConfig:
      LogFormat: JSON
 Resources:
- HelloWorldFunction:
+ ShortenUrlFunction:
    Type: AWS::Serverless::Function # More info about Function Resource: https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#awsserverlessfunction
    Properties:
-     CodeUri: hello_world/
+     CodeUri: shorten_service/
      Handler: app.lambda_handler
      Runtime: python3.13
      Architectures:
@@ -49,13 +49,13 @@ Outputs:
  # ServerlessRestApi is an implicit API created out of Events key under Serverless::Function
  # Find out more about other implicit resources you can reference within SAM
  # https://github.com/awslabs/serverless-application-model/blob/master/docs/internals/generated_resources.rst#api
- HelloWorldApi:
+ ShortenUrlApi:
       Description: API Gateway endpoint URL for Prod stage for Shorten URL function
       Value: !Sub "https://{ServerlessRestApi}.execute-api.${AWS::Region}.amazonaws.com/Prod/shorten/"
- HelloWorldFunction:
-   Description: Hello World Lambda Function ARN
-   Value: !GetAtt HelloWorldFunction.Arn
- HelloWorldFunctionIamRole:
-   Description: Implicit IAM Role created for Hello World function
-   Value: !GetAtt HelloWorldFunctionRole.Arn
+ShortenUrlFunction:
+    Description: Shorten URL Lambda Function ARN
+    Value: !GetAtt ShortenUrlFunction.Arn
+ShortenUrlFunctionIamRole:
+    Description: Implicit IAM Role created for Shorten URL function
+    Value: !GetAtt ShortenUrlFunctionRole.Arn
