+++ tests/unit/test_handler.py
@@ -1,72 +1,37 @@
 import json
-
+import os
 import pytest
+from unittest.mock import MagicMock
 
-from hello_world import app
+# Set the environment variable before importing the app
+os.environ['URL_LOOKUP_TABLE_NAME_DYNAMODB'] = 'mock-table'
 
+# Now we can import from the correctly named folder
+from shorten_service import app
 
 @pytest.fixture()
 def apigw_event():
-    """ Generates API GW Event"""
-
+    """ Generates a sample API GW Event for a POST /shorten request """
     return {
-        "body": '{ "test": "body"}',
-        "resource": "/{proxy+}",
+        "body": '{ "url": "https://www.example.com"}',
         "requestContext": {
-            "resourceId": "123456",
-            "apiId": "1234567890",
-            "resourcePath": "/{proxy+}",
-            "httpMethod": "POST",
-            "requestId": "c6af9ac6-7b61-11e6-9a41-93e8deadbeef",
-            "accountId": "123456789012",
-            "identity": {
-                "apiKey": "",
-                "userArn": "",
-                "cognitoAuthenticationType": "",
-                "caller": "",
-                "userAgent": "Custom User Agent String",
-                "user": "",
-                "cognitoIdentityPoolId": "",
-                "cognitoIdentityId": "",
-                "cognitoAuthenticationProvider": "",
-                "sourceIp": "127.0.0.1",
-                "accountId": "",
-            },
-            "stage": "prod",
+            "stage": "Prod",
+            "domainName": "test.execute-api.us-east-1.amazonaws.com"
         },
-        "queryStringParameters": {"foo": "bar"},
-        "headers": {
-            "Via": "1.1 08f323deadbeefa7af34d5feb414ce27.cloudfront.net (CloudFront)",
-            "Accept-Language": "en-US,en;q=0.8",
-            "CloudFront-Is-Desktop-Viewer": "true",
-            "CloudFront-Is-SmartTV-Viewer": "false",
-            "CloudFront-Is-Mobile-Viewer": "false",
-            "X-Forwarded-For": "127.0.0.1, 127.0.0.2",
-            "CloudFront-Viewer-Country": "US",
-            "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8",
-            "Upgrade-Insecure-Requests": "1",
-            "X-Forwarded-Port": "443",
-            "Host": "1234567890.execute-api.us-east-1.amazonaws.com",
-            "X-Forwarded-Proto": "https",
-            "X-Amz-Cf-Id": "aaaaaaaaaae3VYQb9jd-nvCd-de396Uhbp027Y2JvkCPNLmGJHqlaA==",
-            "CloudFront-Is-Tablet-Viewer": "false",
-            "Cache-Control": "max-age=0",
-            "User-Agent": "Custom User Agent String",
-            "CloudFront-Forwarded-Proto": "https",
-            "Accept-Encoding": "gzip, deflate, sdch",
-        },
-        "pathParameters": {"proxy": "/examplepath"},
         "httpMethod": "POST",
-        "stageVariables": {"baz": "qux"},
-        "path": "/examplepath",
+        "path": "/shorten",
     }
 
-
-def test_lambda_handler(apigw_event):
+def test_lambda_handler(apigw_event, mocker):
+    """ Tests that the lambda_handler successfully processes a valid request """
+    # Mock the DynamoDB table and its put_item method
+    mock_table = MagicMock()
+    mocker.patch.object(app, 'short_url_table', mock_table)
 
     ret = app.lambda_handler(apigw_event, "")
     data = json.loads(ret["body"])
 
     assert ret["statusCode"] == 200
-    assert "message" in ret["body"]
-    assert data["message"] == "hello world"
+    mock_table.put_item.assert_called_once()
+    assert "short_url" in data
+    assert data["long_url"] == "https://www.example.com"
